Refactor Module handling
